LayoutTests/inspector/model/font-calculate-properties-expected.txt

Test for WI.FontStyles.

Basic Properties Test
Feature Property Ligatures Test
Feature Property Position Test
Feature Property Capitals Test
Feature Property Numeric Test
Feature Property Alternates Test
Feature Property East Asian Test
Unregisted Feature Property Test
Variable Font Test

== Running test suite: WI.FontStyles
-- Running test case: WI.FontStyles.BasicPropertiesTest
PASS: Font size property value should be '2rem'.
PASS: Font size property value should be 'italic'.
PASS: Font size property value should be '700'.
PASS: Font size property value should be '150%'.

-- Running test case: WI.FontStyles.FeaturePropertyLigaturesTest
PASS: Property value should be 'no-common-ligatures'.
PASS: 'calt' feature value should be '1'.
PASS: 'dlig' feature value should be '1'.
PASS: 'hlig' feature value should be '1'.

-- Running test case: WI.FontStyles.FeaturePropertyPositionTest
PASS: Property value should be 'super'.
PASS: 'sups' feature value should be '0'.
PASS: 'subs' feature value should be '1'.

-- Running test case: WI.FontStyles.FeaturePropertyCapitalsTest
PASS: Property value should be 'all-small-caps'.
PASS: 'c2sc' feature value should be '1'.
PASS: 'smcp' feature value should be '1'.

-- Running test case: WI.FontStyles.FeaturePropertyNumericTest
PASS: Property value should be 'lining-nums tabular-nums stacked-fractions'.
PASS: 'zero' feature value should be '1'.

-- Running test case: WI.FontStyles.FeaturePropertyAlternatesTest
PASS: Property value should be 'historical-forms'.
PASS: 'hist' feature value should be '1'.

-- Running test case: WI.FontStyles.FeaturePropertyEastAsianTest
PASS: Property value should be 'jis78 full-width'.
PASS: 'jp78' feature value should be '1'.

-- Running test case: WI.FontStyles.UnregisteredFeaturePropertyTest
PASS: 'appl' feature value should be '1'.

-- Running test setup.
PASS: Font should be loaded.
-- Running test case: WI.FontStyles.VariableFontTest
PASS: Width property value should be '50%'.
PASS: Width axis name should be 'Width'.
PASS: Width axis value should be '777'.
PASS: Width axis minimum should be '0'.
PASS: Width axis maximum should be '1000'.
PASS: Width axis default should be '750'.
PASS: Descent axis name should be 'Descent'.
PASS: Descent axis value should be '100'.
PASS: Descent axis minimum should be '0'.
PASS: Descent axis maximum should be '250'.
PASS: Descent axis default should be '0'.
PASS: Height axis name should be 'Height'.
PASS: Height axis value should be 'undefined'.
PASS: Height axis minimum should be '0'.
PASS: Height axis maximum should be '1000'.
PASS: Height axis default should be '750'.
PASS: Tracking axis name should be 'Tracking'.
PASS: Descent axis value should be '500'.
PASS: Tracking axis minimum should be '0'.
PASS: Tracking axis maximum should be '1000'.
PASS: Tracking axis default should be '0'.
